endpoint getProfile
